+++ src/types/pluralize.ts
@@ -1,0 +1,5 @@
+export type PluralizeWordsType = {
+  one: string;
+  two: string;
+  five: string;
+};